Move component ScrollableChart from 'components' to 'components/index/_shared/charts/_base'
