+++ components/index/CardsMonitoring/HospitalizedNumber/Chart.vue
@@ -69,7 +69,7 @@ import DataViewTable, {
   TableHeader,
   TableItem,
 } from '@/components/DataViewTable.vue'
-import ScrollableChart from '@/components/ScrollableChart.vue'
+import ScrollableChart from '@/components/index/_shared/charts/_base/ScrollableChart.vue'
 import {
   DataSets,
   DataSetsPoint,
