+++ components/index/CardsMonitoring/PositiveRate/Chart.vue
@@ -114,7 +114,7 @@ import DataViewTable, {
   TableHeader,
   TableItem,
 } from '@/components/DataViewTable.vue'
-import ScrollableChart from '@/components/ScrollableChart.vue'
+import ScrollableChart from '@/components/index/_shared/charts/_base/ScrollableChart.vue'
 import {
   DisplayData,
   yAxesBgPlugin,
