+++ components/index/CardsMonitoring/UntrackedRate/Chart.vue
@@ -117,7 +117,7 @@ import DataViewTable, {
   TableHeader,
   TableItem,
 } from '@/components/DataViewTable.vue'
-import ScrollableChart from '@/components/ScrollableChart.vue'
+import ScrollableChart from '@/components/index/_shared/charts/_base/ScrollableChart.vue'
 import {
   DisplayData,
   yAxesBgPlugin,
