+++ components/index/_shared/charts/_base/ScrollableChart.vue
